==> R/00_all.qmd <==
---
title: "Final project: Group 1"
format:
  html:
    embed-resources: true
editor: visual
---

# Members

-   Anna Lifousi (`s232979`)
-   Jordan Sylvester Fernandes (`s222497`)
-   Manuel Arcieri (`s230158`)
-   Quim Bech Vilaseca (`s233374`)
-   Xavier Viñas Margalef (`s233532`)

{{< include 01_load.qmd >}}

{{< include 02_clean.qmd >}}

{{< include 03_augment.qmd >}}

{{< include 04_describe.qmd >}}

{{< include 05_analysis.qmd >}}
