==> R/05_analysis.qmd <==
---
title: "05_analysis.qmd"
format:
  html:
    embed-resources: true
editor: visual
---

## Analyse data

```{r}
data <- read_tsv(str_c(data_dir, '03_dat_aug.tsv.gz'), col_types = cols(
  Pregnancies = col_integer(),
  Glucose = col_integer(),
  BP = col_integer(),
  Skin_thick = col_integer(),
  Insulin = col_integer(),
  BMI = col_number(),
  DPF = col_number(),
  Age_onset = col_integer(),
  Outcome = col_factor()
  ))
```

### Boxplots

```{r}
#| warning: false
Boxplots <- function(attribute) {
    Boxplot <- ggplot(data, aes(x = 1, y = data[[attribute]])) +
               geom_boxplot() +
               theme_minimal() +
               labs(x = NULL, y = attribute)
    return(Boxplot)
}

Pregnancies <- Boxplots("Pregnancies") +
    labs(title = paste("Boxplots of all the attributes"))

Glucose <- Boxplots("Glucose")

BP <- Boxplots("BP")+
    labs(y = "Blood Pressure")

Skin_thick <- Boxplots("Skin_thick") +
    labs(y = "Skin Thickness")

Insulin <- Boxplots("Insulin")

BMI <- Boxplots("BMI")+
    labs(y = "Body Mass Index")

DPF <- Boxplots("DPF") +
    labs(y = "Diabetes Ped. Func.")

Age_onset <- Boxplots("Age_onset") +
    labs(y = "Age Onset")

combination_plot <- (Pregnancies + Glucose + BP + Skin_thick + Insulin + BMI + DPF + Age_onset) +
    plot_layout(guides = 'collect', ncol = 4)

ggsave("../results/boxplot.png", bg = "white")

combination_plot
```

### Violin Plots

```{r}
#| warning: false
Violinplots <- function(attribute) {
    Violinplot <- ggplot(data, aes(x = 1, y = data[[attribute]])) +
               geom_violin(fill = "#03A9F4", alpha = 0.5) +
               theme_minimal() +
               labs(x = NULL, y = attribute) + 
               geom_boxplot(width = 0.2) + 
               stat_summary(fun.y = mean, geom = "point",
                            size = 2, color = "red")
    return(Violinplot)
}

Pregnancies <- Violinplots("Pregnancies") +
    labs(title = paste("Violin plots of all the attributes"))

Glucose <- Violinplots("Glucose")

BP <- Violinplots("BP")+
    labs(y = "Blood Pressure")

Skin_thick <- Violinplots("Skin_thick") +
    labs(y = "Skin Thickness")

Insulin <- Violinplots("Insulin")

BMI <- Violinplots("BMI")+
    labs(y = "Body Mass Index")

DPF <- Violinplots("DPF") +
    labs(y = "Diabetes Ped. Func.")

Age_onset <- Violinplots("Age_onset") +
    labs(y = "Age Onset")

combination_plot <- (Pregnancies + Glucose + BP + Skin_thick + Insulin + BMI + DPF + Age_onset) +
    plot_layout(guides = 'collect', ncol = 4) 

combination_plot
ggsave("../results/violin_plot.png", bg = "white")

```

### Add "age_ranges" column with the age of onset values in ranges in intervals of 10

```{r}
age_ranges <- c(18, 29, 39, 49, 59, 69, 79, 89, 99)

age_range_labels <- c("18-29", "30-39", "40-49", "50-59", "60-69", "70-79", "80-89", "90-100")

data <- data |> 
    mutate(Age_range = cut(Age_onset, breaks = age_ranges, labels = age_range_labels))

head(data)
```

### Age ranges density plot.

```{r}
density_age <- ggplot(data, aes(y = Age_range)) +
    geom_bar(fill = "#7a5195") +
    theme_minimal() +
    labs(x = "Age Range", y = "Count", title = "Age distribution in the dataset")+
    coord_flip()

density_age
ggsave("../results/density_age.png", bg = "white")

```

### Correlation Heatmap

```{r}
data_clean_from_nas <- data |>
  drop_na() |>
  select(where(is.numeric))


correlation_matrix <- cor(data_clean_from_nas)


correlation_data <- as.data.frame(as.table(correlation_matrix))
colnames(correlation_data) <- c("Var1", "Var2", "Correlation")


heatmap <- ggplot(correlation_data, aes(Var1, Var2, fill = Correlation,
                                        label = sprintf("%.2f", Correlation))) +
  geom_tile(color = "white") +
  geom_text(color = "black", size = 3) +
  scale_fill_gradientn(colors = colorRampPalette(c("#ffffff", "#ef5675", "#7a5195"))(200),
                       na.value = "white") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(title = "Correlation Heatmap")
ggsave("../results/heatmap.png", bg = "white")

heatmap
```

### Linear Regression: Skin Thickness vs Blood Pressure

```{r}
#| warning: false
model <- lm(`Skin_thick` ~ BP, data = data)

ggplot(data, aes(x = BP, y = Skin_thick)) +
  geom_point(color = "steelblue", alpha = 0.7, size = 3) +  
  geom_smooth(method = "lm", se = FALSE, color = "red", linewidth = 1.5, formula = 'y ~ x') + 
  labs(x = "Blood Pressure", y = "Skin Thickness", 
       title = "Linear Regression: Skin Thickness vs Blood Pressure",
       subtitle = "Relationship between Skin Thickness and Blood Pressure") +
  theme_minimal() +  
  theme(plot.title = element_text(size = 14, face = "bold"),
        plot.subtitle = element_text(size = 10, color = "black")) +
  guides(color = guide_legend(title = NULL))
ggsave("../results/linear_reg.png", bg = "white")

```

### Principal component analysis

```{r}
pca_data <- data |>
    drop_na()

pca_model <- pca_data |>
    select(where(is.numeric)) |>
    prcomp(scale. = TRUE)

pca_model |>
    augment(pca_data) |>
    ggplot(aes(.fittedPC1, .fittedPC2, color = Outcome)) +
        geom_point(size = 1.8, alpha = 0.6) +
        theme_minimal() +
        scale_colour_manual(values = c("#F44336", "#03A9F4")) +
        labs(x = "PCA component 1", y = "PCA component 2",
             title = "Principal component analysis") + 
        theme(plot.title = element_text(hjust = 0.5))
ggsave("../results/PCA.png", bg = "white", width = 10, height = 5)

```

### Variance explained

Goal: predict diabetic or non-diabetic. PCA : with the outcome variable.

```{r}
pca_model |>
    tidy(matrix = "eigenvalues") |>
    ggplot(aes(PC, percent)) +
        geom_col(width = 0.5, fill = "#03A9F4") +
        scale_x_continuous(breaks = 1:8) +
        scale_y_continuous(labels = scales::percent_format()) +
        theme_minimal() +
        labs(x = "Principal components", y = "Variance explained",
             title = "Variance explained by each principal component") + 
        theme(plot.title = element_text(hjust = 0.5))
ggsave("../results/PCA_Var.png", bg = "white", width = 10, height = 4)

```

### BMI class distribution

```{r}

rows <- nrow(data)
id <- seq(1, rows, by = 1)

pivot_data <- data |>
  select(BMI_class, BMI) |>
  mutate(id) |>
  pivot_wider(names_from = BMI_class,
              values_from = BMI,
              id_cols = id) |>
  select(-id, -"NA")

non_na_counts <- colSums(!is.na(pivot_data[, !(names(pivot_data) %in% c("id", "NA"))]))

pivot_data_plot <- data.frame(columns = names(non_na_counts), count = non_na_counts)

ggplot(pivot_data_plot, aes(x = columns, y = count, fill = columns)) +
  geom_bar(stat = "identity", position = "dodge",
           width = 0.7, color = "black") +
  labs(
      title = "Number of diabetes cases in different BMI class",
      y = "Count"
  ) +
  theme_minimal() 

head(pivot_data)
ggsave("../results/BMI_class.png", bg = "white")
```
